Moved cs_process functions to main.py
